Migration add field existing checker

+++ database/migrations/2016_08_23_052001_add_reference_id_column_to_users_table.php
@@ -2,6 +2,7 @@
 
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Database\Migrations\Migration;
+use Illuminate\Support\Facades\Schema;
 
 class AddReferenceIdColumnToUsersTable extends Migration
 {
@@ -12,10 +13,12 @@ class AddReferenceIdColumnToUsersTable extends Migration
      */
     public function up()
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->integer('reference_id')->unsigned()->index()->nullable()->after('api_token');
-            $table->foreign('reference_id')->references('id')->on('users')->onDelete('set null');
-        });
+        if(!Schema::hasColumn('users', 'reference_id')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->integer('reference_id')->unsigned()->index()->nullable()->after('api_token');
+                $table->foreign('reference_id')->references('id')->on('users')->onDelete('set null');
+            });
+        }
     }
 
     /**
